tab bar factory, and fixing constraints in onboarding page view
